backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0

langchain==0.2.14
langchain-community==0.2.12
langchain-text-splitters==0.2.2
langchain-qdrant==0.1.4

qdrant-client==1.11.1
pypdf==4.3.1

mlflow==2.14.3
httpx==0.27.0
google-genai==0.6.0
